+++ test/engine/fun.js
@@ -15,8 +15,12 @@ describe('fun', function () {
         '6': 'return',
         '7': '// this is a comment',
         // The following tasks do not contribute to result.
-        '8': "'This is a string with a.dot'",
-        '9': '1.2'
+        // They are added to test there are no conflicts with
+        // global tasks resolution.
+        'a': "'This is a string with a.dot'",
+        'b': '// This is a comment with a.dot',
+        'c': '&Math.cos',
+        'd': '1.2'
       },
       pipe: {
         'a': [ '0', '2', 0 ],